fix bad imports from parse-shortcode.ts
