+++ src/core/lib/parse-shortcode-types.ts
@@ -1,0 +1,15 @@
+/*
+ * parse-shortcode-types.ts
+ *
+ * Copyright (C) 2023 Posit Software, PBC
+ */
+
+export interface Shortcode {
+  name: string;
+  rawParams: {
+    name?: string;
+    value: string;
+  }[];
+  namedParams: Record<string, string>;
+  params: string[];
+}
